new class Figure
